Update index.html
Updated `LinguisticObject` `value` property to `content` property, and updated first mention of "the Linguistic Object class" under the "Statements as a Resource" section to "the `LinguisticObject` class" as per other class type references in the documentation. It was unclear if it would be appropriate to modify the second plural reference to Linguistic Objects in the same paragraph?